// src/app/models/businessUser.ts
export class BusinessUser{
    cmsys_user_id: number = 0;
    company_name: string = '';
    owner: string = '';
    address: string = '';
    city: string = '';
    state: string = '';
    zipcode: string = '';
    country: string = '';
    email: string = '';
    mobile_number: string = '';
    tag: string = '';
    industry: string = '';
    website: string = '';
    company_description: string = '';
    password: string = '';
    user_otp: string = '';
}